npm shrinkwrap
This will ensure that the versions of the dependencies of the dependencies
remain the same on npm install, that way we can ensure bitcore works as
intended for the end-user. Note that this does not ensure byte-for-byte
compatibility. We may address that issue in the future.

See: https://www.npmjs.org/doc/cli/npm-shrinkwrap.html