support the case where r, s are invalid
Rarely it will be the case that r, s are invalid when using a deterministic k.
RFC 6979 specifies how to handle this case by iterating through new values of k
until a suitable one is found. This allows one to pass the badrs value to
deterministicK(badrs) so that you can iterate through more values of k in case
r, s are bad. This is used in the signing function; normally one ought not need
call deterministicK directly.