change .sign and .verify to return this
One of the primary goals of fullnode as a library is internal consistency, and
part of the philosophy of the objects is that the instance (protoype) methods
modify the object and then return this. This allows you to change the functions
together, e.g. ecdsa.sign().verify().sign(). Although this may or may not be
useful in particular for signing and verifying, in order to retain the internal
consistency and be like the other objects, ECDSA should behave that way.