add fromX prototype convenience method
The way objects in fullnode works is that typically you create a blank new
object, and then if you want to "set" something, you set an object. In this
case:

var p = Point().fromX(isOdd, x);

That's how it should work. Since we are now committing to simply extending
(rather than inheriting) the BN and Point classes from bn.js and elliptic, we
can easily create a prototype method for fromX by just setting all the
enumerable properties (the prototypes should be the same). So long as there are
no non-enumerable properties of a point, this should work correctly.