do not override prototypes
When creating a new point with prototype.fromX, we should not copy the
prototype properties, because they are already set in the prototype.

Also add a test that makes sure this code is resilient against future changes
of elliptic's Point class. I wasn't really sure what the elements of elliptic's
Point class were, so here we make sure that the values they are today ['curve',
'type', 'precomputed', 'x', 'y', 'inf'] stay that way in any future updates of
this code. If the internal values of a Point ever change in elliptic, we will
have to thoroughly understand what they are and possibly update the fromX
function to act accordingly, if there is ever a discrepancy about what is
enumerable.