add reddit to copyright notices
The history of this software is odd. It was originally written by myself, based
on bitcore and BitcoinJS, while I worked for BitPay. I then left BitPay, and
during the brief period when I did not work for anybody, I forked it and called
it fullnode. Now I work for reddit. I think it is fair to give copyright to
reddit for all the new stuff I create. However, all the other copyright notices
should remain, including the copyright notice for myself.