nhashtype is actually last byte, not last 4 bytes
A signature in a script in a transaction has one byte appended to the end which
indicates the hash type (e.g., SIGHASH_ALL, SIGHASH_SINGLE, etc.). There was a
bug in my signature encoding code that treated this as 4 bytes. That is
incorrect. It should be a single byte.