CScriptNum
bitcoind has a class called CScriptNum which is used to handle numbers in the
stack when a script is executed. These are interpreted as 4 byte, SM-encoded
little endian numbers. There is important feature that distinguishes them from
normal 4 byte numbers - they can overflow. Thus, internally, they are
represented as 8 byte numbers. When overflowed and output, the overflowed value
is output. Thus the output can be longer than 4 bytes. However, you can not
perform operations on numbers over 4 bytes (nMaxNumSize). If you do, an
exception is thrown. Thus, if an overflow occurs and you output an overflowed
number, that's fine, but if you try to operate on it again, that's not fine.

I have implemented this in fullnode by adding the fromCScriptNumBuffer and
toCScriptNumBuffer functions. These will be used in the interpreter.