use to/from CScriptNumBuffer
If the number being input is greater than 4 bytes, an exception should be
thrown. This is explained in the CScriptNum comments in bitcoind. As explained
in the previous commit, I have added this functionality to fullnode in the BN
class by adding to/from CScriptNumBuffer functions. Here I simply update the
interpreter to use them. This should increase the correctness of the
interpreter. For instance, if a script tries to perform a number operation on a
5 byte value, that will now throw an error, as it should.