fix CHECKSIG by correcting endiannes of hashbuf
bitcoin assumes hash is little endian before converting it to the e value used
in ECDSA. I had been using big endian, and this is why the test was broken.
Since the ECDSA class is specific to bitcoin, I should probably update the
default from big endian to little endian. But for now, the default is still big
endian, and the option of little endian is available for signing/verifying
transactions.