NULLDUMMY
An optional flag when verifying scripts is that the extra push data for a
multisig must be a "null dummy", i.e. a blank value. That is because the bug in
OP_CHECKMULTISIG that pops one too many items from the stack, and this extra
value is a source of transaction malleability, unless it is forced to be blank.