pass all script_valid and script_invalid tests
* Fix MAX_SCRIPT_SIZE typo
* Add SCRIPT_VERIFY_DISCOURAGE_UPGRADABLE_NOPS flag from latest bitcoind
* Add checkSignatureEncoding
* Add checkPubkeyEncoding
* Use len vs. length in script chunks to throw error - turns out that was
indeed an issue after all
* Use all flags in tests
* Other minor fixes and additions

...and for the record, here is some information on what constitutes "hybrid pubkeys":

* http://sourceforge.net/p/bitcoin/mailman/message/29416133/

Test coverage is also significantly improved compared to the last time that I
ran coverage. That is because the step() function of the interpreter is very
large, and went from being poorly tested to well tested. Overall, the
"branches" coverage is now higher, because of course the step() function is one
giant switch/case block.
=============================== Coverage summary ===============================
Statements   : 93.89% ( 3088/3289 )
Branches     : 83.73% ( 1287/1537 )
Functions    : 97.07% ( 365/376 )
Lines        : 93.88% ( 3069/3269 )
================================================================================