pass all but one relevant tx_valid, tx_invalid tests
Added tests to the script interpreter for all relevant tx_valid and tx_invalid
tests from bitcoin core. Note that most of these tests are relevant, but not
all are.  Some of the tests rely on checks that should not be performed by the
script interpreter, particularly those that rely on checks from
CheckTransaction in bitcoind. Those tests have been disabled. There is one
other test from tx_invalid that tests some combination of SIGHASH_ALL and
SIGHASH_ANOYONECANPAY which fails, and I'm fairly sure it should not fail. I
will fix that soon. The changes I've made to pass these tests are:

* CheckSignatureEncoding - does not decode signature unless flags are set
* findAndDelete - now also compare pushdata opcode (not sure if this is exactly
 equal to bitcoind code)
* catch exceptions around parsing signatures and pubkeys in CHECKSIG and
 CHECKMULTISIG