add comments to move the tx_valid/invalid tests
The tx_valid and tx_invalid tests are being run in the script interpreter
tests, however they really rely on other checks that go beyond the script
interpreter. The tx_invalid checks, for instance, rely on the CheckTransaction
function in bitcoind, which I have not yet translated to javascript. Once I
create a txbuilder class, that is probably where the CheckTransaction
translation will go, and therefore these tests should really be moved to the
tests for the txbuilder. Once that happens, I can enabled the small handful of
disabled tests that rely on CheckTransaction.