add bitcoin core copyright to license
...and rearrange copyrights so that the most important ones are first. Since
this library at one point morphed into bitcore2, I place bitcore as the most
important extra copyright to indicate the importance of BitPay funding 4 or so
weeks of development of this project. Second is BitcoinJS, where some important
code comes from, such as BIP32 and public key recovery from a signature. Then
comes bitcoind, where the script interpreter is translated from. Finally is
BitcoinJ, which I have not personally referenced at all, but which is here
because it was in the original copyright for BitcoinJS.

The top copyright goes to myself and reddit. I own the trademark "fullnode"
since I created that name while independent and not working for anyone.
However, reddit owns the copyright on the code I am writing while under their
employment.