add tx verification to txbuilder
The transaction builder is both used to build a transaction, and also to verify
that a built transaction is valid. This ultimately will enable sharing
transactions with others off-blockchain and verifying that they are valid. Of
course it could also ultimately be used to validate transactions in the
blockchain.

The tests that had been checking if a transaction are valid in the script
interpreter tests have been moved to the txbuilder tests. This is where the
tests belong. The script interpreter tests only run script_valid and
script_invalid. The tx_valid and tx_invalid tests should also run the
CheckTransaction tests. Those are now actually being executed as the checktx
function now exists and the txbuilder contains that verificaion logic plus the
script interpreter verification.