change name: txidbuf -> txhashbuf
The hash in an input is actually just a plain old hash. The id would be the
reverse of the hash. Since the thing in an input is *not* reversed, it is just
the hash, not the id. This merely changes the name to accurately reflect what
the thing is - it is a hash of a transaction. Not an id.