downrade chai from 1.10.0 to faster 1.9.0
On the "new" version of chai, 1.10.0, te browser tests take about 30s on my
machine. On the "old" version 1.9.0, the tests take about 25s. So the tests are
substantially faster with the old version of chai. Since I have to run the
tests all the time, I prefer the older version, which is faster (I have no idea
what new things 1.10.0 brings to the table, but speed isn't one of them). This
also makes me want to remove chai altogether - it is mostly useful for our
purposes.