make aes, cbc, aescbc first-class
I have moved the AES, CBC, and AESCBC classes out of the "expmt" folder and
into the main folder. These classes will be combined with an HMAC in a new
class that will probably be called AESCBCHMAC, and which will be used to
encrypt the wallet files that don't yet exist. The wallet class will probably
be called "Wallet" and it will need to have a way to export it to an encrypted
form and imported it from an encrypted form.

For the record, it is often argued that encryption and decryption should use
existing libraries. In our case, that would be SJCL. That is a fine choice much
of the time, except here we only need a tiny fraction of SJCL, and much
functionality from that library is duplicated in fullnode already. I think it
is better not to add an extra dependency, when the cryptographic primitives are
already available, or easily imported.

For instance, CBC is very easy to implement. It would be inappropriate to
include SJCL simply for CBC. AES, on the other hand, is easily imported from
the AES package without including all the extra baggage that comes with SJCL.
And furthermore, the HMAC we will ultimately need for encrypting/decrypting
data is also already contained in fullnode for other reasons.

Ultimately, there should be a standard wallet format, and a standard way to
encrypt and decrypt that wallet format. fullnode aims to help design such a
standard, and have accurate implementations of the cryptography for encrypting
and decrypting such standard wallets.