use hash.js for sha512 and other sha-2 functions
The latest version of hash.js now includes sha512. Since we already have to
include hash.js for ripemd160, and because it is a dependency of elliptic, it
simplifies our code somewhat to use hash.js for all hash functions.
Unfortunately the version of elliptic we are using has an outdated hash.js
dependency, so we are now including two different versions of hash.js. That
will be fixed in a future version of elliptic.