add index.html for easy testing, and fix standalone
I've added an index.html and index-min.html that load fullnode.js and
fullnode-min.js, respectively, for easily testing fullnode in the browser
console. browserify and es6ify don't seem to get alone well when trying to
specify a "standlone", i.e. setting the fullnode global variable, so the way I
have set this is by setting global.fullnode in index.js. It's quite possible
the standalone was never actually working before, as I had never tested it. But
in any case, it works now.