move mainnet/testnet objects to injector pattern
Rather than pass around networkstr = "testnet" or "mainnet" everywhere, it's
convenient to be able to inject dependencies into the classes, including what
constants to use. Furthermore, testnet addresses should actually be invalid on
mainnet, and vice versa. I've invented a new injector pattern for this purpose.
If this works well, I may convert the entire library to use this pattern. Not
only will it be useful for overriding constants, but entire dependencies, for
instance for overriding the (yet to be added) database for testing purposes.

I've investigated a few libraries for doing this, notably proxyquire, but it's
pretty easy to do this without including any more libraries which I would
rather not have to audit.

The way the injector pattern works is that classes that need to have
dependencies overrided, particularly Address, Privkey and BIP32 (for now), they
keep track of which dependencies have been injected via the classmap. If you
request a new class with a new dependencies object, it will make a new one,
otherwise it will return the one it already created. There is one gotcha - if
you want to use the same exact class and not create a new one each time, you
need to use the same exact dependency object (not a clone). This will probably
trip people up. But it works for now.