remove fasteq (use Buffer.compare); speed up cmp
Since fasteq is now made irrelevant by Buffer.compare, which is built in to
io.js, we don't need it anymore (it wasn't being used anyway). Thus I've
updated cmp to be equal to just the constant-time algorithm. Furthermore, I've
made a minor speed improvement. Rather than use arithmetic, I use an XOR, which
is theoretically faster (although I haven't tested it). I also almost
introduced a bug while changing the algorithm that was caught by some unrelated
tests. I've added a new test to ensure that particular bug never occurs (if the
buffers were not equal in length, I assumed the remaining values were 0, which
is certainly an error, which might lead to some buffers being considered equal
when they are not).