move BIP39 to own dir; convert to inject pattern
In order to make it possible for people to include only the English or Japanese
or other word lists, I'm moving BIP39 to its own folder along with two new
files for the wordlists, en.js and jp.js, and migrating to the injector
pattern.  Both wordlists are still included, but now it's possible for people
to override them with other wordlists externally. I will probably ultimately
remove the Japanese wordlist from the defaults as well.

I've also migrated toMnemonic/fromMnemonic to toString/fromString which is more
idiomatic.