switch injector to memoize
There was an issue with the injector, which was that it was impossible to use
the injector to get back a cached class from outside of the file that created
it, because "dependencies" was not being exposed. Rather than attach the
dependencies object to every class to make that possible, I've changed the
inject pattern to use memoization. Now the .inject() method that is exposed by
every class using this pattern is memoized. That way you can make a new class
with equivalent injected dependencies without having to worry about creating a
bunch of new classes, wasting memory.