migrate entire library to use inject pattern
The inject pattern is where a class lets you override its dependencies. This is
useful both to distinguish between mainnet and testnet, where we want to
override constants, and also to distinguish between node and the browser, where
we will (not yet) have a database and network connections that have completely
different dependencies. It is also useful for testing, where, when desired, we
will be able to override the database with a testing-only in-memory database.

The way inject works is that each class simply provides an inject function that
lets you specify its dependencies. The injector class is a convenience class
for caching the created classes, so that if you create a class with the same
dependencies, you get back the same class. In practice, the injector class
probably only needs to be used internally, and a class can have its
dependencies overridden by calling Class.inject(deps).

Note that there is one tricky property of the inject pattern discussed in
earlier commits. In order to get a cached class, you must inject the exact same
dependencies object. Not just the same list of dependencies - but the same
object. If you create a class with a different dependencies object, you will
get a new class. If you want to make sure not to abuse memory and get back the
same class, you need to use the same object.

In other words:

var deps = {};
Address.inject(deps) === Address.inject(deps);

but,

Address.inject({}) !== Address.inject({});

...because in the latter, two different objects are actually being created.
This is a property of the ES6 Map that we are using for the injector. This
might be fixed in a future version of the injector, but not now.