Work and Workers
As we move towards supporting the bitcoin p2p protocol, it is obvious
that pure javascript cryptography is slow enough that it will block IO.
Even in node, where we will be able to add support for C or C++ code for
the cryptography, the fact that other full node implementations support
threaded cryptography indicates we ought to do the same thing. Thus I've
added the Work and Workers classes to manage this.

In the browser, there is one way to do this: web workers. Web workers
allow one to create threads and pass messages to and from them. The
worker threads do not block the IO of the main thread. In addition, it
is even possible to transfer, not copy, data to them, to save on the
amount of memory that has to be copied. I have implemented support for
web workers in the new Workers class that manages the pool of workers.
Right now Workers only supports one thread, but now that the basic
interface is in place, I will add support for more threads in the
future. It also does not yet support memory transfer, only copy, but
that will also be relatively easy to add later.

Node, unfortunately, does not have native support for threads. There are
two libraries I investigated using: webworker-threads and
threads-a-gogo. Unfortunately both of these are not very suitable
because they do not work with node's require(). That is, it is not
possible to require modules inside the threads. It is possible to
include pure javascript code via importScripts, but that would prevent
us from using faster C or C++ code in the future. Thus, I have opted not
to use threads in node, and instead have chosen to use node's built-in
child_process fork. This allows creating new process forks that
communicate by message passing. This will be slower than transferring
data in a web worker, but that seems to be the best I can do in node
right now.

The Workers class is used by Work, a convenient class for computing
units of work. The way Work works is that you give it an object and then
have it compute a method on that object and then return the result. The
computation is done by the workers. Because most fullnode objects
already work this way, this is a very suitable way to manage work. The
very first class that I will use this with is Connection, which will
need to validate Msg objects received from the network. The validation
includes checking a hash, and is thus CPU-intensive. Obviously, after
Msg, the next steps will include transaction and block validation, which
will be far more computationally intense, as it involves running the
script interpreter with lots of signature verifications.

Besides adding Work and Workers, I have also had to make several other
updates to fullnode in order to make this possible. Particularly, I have
had to edit the way the browser bundles are built so that karma and
mocha can both understand the web workers correctly. Since web workers
have to load an "outside" script via importScripts, there is a new file
being produce called fullnode-worker.js (or fullnode-worker-min.js).
That is the web worker, which must in turn load fullnode.js in order to
access library. I have also added a small express app to view the mocha
tests (if choosing not to run karma), since loading the HTML pages from
the filesystem has cross-origin policy issues with the importScripts
method used by the workers.