await pattern
Since I'm experimenting with the use of promises and generators to
handle asynchronous stuff in fullnode, I can't use established notions
like streams. Instead I have to find new ways to do things. The notion
of a generator that yields promises is a general way to deal with
asynchronous stuff, but I don't yet have good names for my methods that
are consistent with the rest of fullnode. It has occurred to me the
perfect word for some of these functions is "await". Whenever a
generator yields promises to the same type of thing over and over again,
like buffers, Msgs, Connections, or whatever, it can't be that we are
".getMsgs()", because ".get" is always used in a synchronous manner.  But
if we are ".awaitMsgs()", that's perfect, as the name implies the
asynchronous nature of the method.