update dependencies
I've updated everything besides bs58. The reason to not updated bs58 is that
the 1.x version of bs58 used Buffers, and the 2.x version used ArrayBuffers.
I'm not sure what the 3.x version uses, but it doesn't seem to work with
fullnode, and almost certainly doesn't fix any fundamental bugs (because the
algorithm used is not error-prone and passes a lot of tests in fullnode).