js standard nits - reduce from 2308 to 1723
I've fixed the js standard nits (what you get when running "standard" on
fullnode) for the files in lib/* and test/* alphabetically up to br.js. All the
nits are very minor - only one of any substance was an odd checking of a string
'0' to 0 and relying on javascript's loose == operator to confirm them as
equal.