remove unused blocksize variable
I've applied js standard to lib/cbc.js and test/cbc.js, and discovered a minor
error, which was that the blocksize value was not actually being used in a
couple of cases. It turns out not to be necessary in those cases, so I've
simply removed them, and also fixed other js standard nits.