Script: Account for reverseMap name inconsistencies
There are a few cases where the opcode name doesn't match the name exactly.
These are mostly related with data pushes, and the cases for `OP_0` and `OP_1NEGATE` were
not handled as data pushes as the buf.length was 0. This adds these exceptions to the
`_chunkToString` method on Script.

References:
- https://github.com/bitcoin/bitcoin/blob/e54ebbf6009716a7abcd4d8d3f7bd910e88decdc/src/script/script.cpp#L13
- https://github.com/bitcoin/bitcoin/blob/e54ebbf6009716a7abcd4d8d3f7bd910e88decdc/src/core_write.cpp#L75