babel-preset-react
fullnode does not use react, but datt does, and datt was having a problem
building fullnode when linking locally since it was looking for the react
preset inside fullnode and wasn't able to follow the symlink back to datt. I
think this is a bug with babel, not datt or fullnode. In any case, the easiest
thing to do was at babel-preset-react as a devDependency of fullnode to fix the
problem.