struct toFastBuffer/fromFastBuffer
It is sometimes the case that an object's toBuffer and fromBuffer use slow
cryptography operations and thus are blocking if used in the main thread. The
idea of toFastBuffer and fromFastBuffer are to have alternative methods for
some classes where this is the case. Particularly BIP32, and other related
classes that use hash functions or ECDSA. I will be adding toFastBuffer and
fromFastBuffer methods to these classes to serialize to and from buffers much
more quickly than with blocking cryptography. The intent is to use the
toFastBuffer and fromFastBuffer to send objects from one thread to another (or
one process to another) where the real heavy lifting can be done in any thread
other than the main one.