use asink for some network stuff
I'm beginning the process of rewriting the networking code for fullnode. The
networking code was created last summer partially as an attempt for me to learn
how best to use native promises in ES6. I have since learned that promises are
only appropriate for "one time" asynchronous things, such as retrieving a
database record. Retrieving a database record is asynchronous, and occurs only
once - and it either succeeds or fails. But for networking where messages are
being sent, a message may be received many times. Promises are not appropriate
in that case. The networking code is thus better off using events or streams.
Fortunately there is not much networking code, and it should be straightforward
to rewrite it to use events or streams. This update only begins the process of
replacing the networking code. I will need to spend a bit more time reminding
myself how my own code works before I can go about replacing it, to be sure I
don't break any working functionality in the process.