struct.toFastBuffer understands blank objects
It will often be the case that I want to send a "blank" object to a worker, for
instance for Address().fromPubkey(pubkey). We want to send the keyless, or
blank, object Address() to the worker. Now Struct natively converts blank
objects to/from 0-length buffers to support this common use-case.