temporarily add code i used to find big block
Oddly, I couldn't find a blockchain API that simply returned a raw block in hex
form. Even Bitcoin Core seems unable to return a plain block. All APIs seem to
return a parsed block only. This means there was no easy way for me to test my
block hashing code. First, I needed to produce an actual plain hex block, and
then try hashing it. This is the code I used to do that, using the output of a
block from Bitcoin Core. I am now going to remove this code since I don't want
to support it and I don't want people thinking it is an example to refer to
about how to use fullnode, but I wanted it to be in here just so I could
personally reference it myself again later.