test can asynchronously hash large data
Although there is already a test in place to test hashing a 1mb block, I
figured it would also be good to test that we can hash larger amounts of data,
in this case a 5mb value. I test both that we can test it asynchronously, and
that the asynchronous value ends up being the same as the synchronous value.