txbuilder async sign
Signing an input is a blocking operation, so it should be put in a worker.
There is one difference between this new method and most other "async" methods
that put stuff in workers, which is we are not simply calling a synchronous
method inside a worker, but rather reimplementing some of the same code in the
asynchronous version. The reason for this is that a Txbuilder object is a large
object, and we don't want to send the entire thing to the worker. All we want
to do is sign the one input, which requires sending a transaction to the
worker, but not all the other properties of a Txbuilder object (such as all
unspent txouts). This is actually one of the primary reasons why the new worker
interface works the way it does - only a Txbuilder object knows how best to
send its own data to a worker to perform an operation.