remove networking & stealth from library
The networking code was never fully working and was poorly architected.
fullnode does indeed need the bitcoin p2p protocol and networking code, but
after what I have learned from the way I implemented the p2p protocol in Datt,
I would substantially change the architecture. Basically, the asynchronous
terminology in fullnode was very hard to understand, and was best rewritten
from scratch.  Furthermore, it was not being used anywhere. So rather than
support the current architecture, it was better removed.

The stealth address functionality was a little different. Basically stealth
addresses as implemented in fullnode never become popular and are not used in
any wallets. Instead, a standard called "reusable payment codes" in BIP 47 has
much more traction. However, there is no imminent need to have either of these
in fullnode, so rather than support the unusable stealth address implementation
in fullnode, it is better to remove it altogether.