MerkleNode
Merkle trees are a core principle in bitcoin which are primarily used to easily
determine is a transaction is in a block without necessarily possessing the
entire block. They are also one of the most importnat things, besides the p2p
protocol, yet to be implemented in fullnode. A MerkleNode has a hash and links
to two subnodes.