fullnode -> Fullnode
Normally, a "class" will be capitalized. Effectively, the main Fullnode object
is a class, and therefore should be capitalized. At some future date, we will
finish implementing the bitcoin p2p protocl, and you will want to make an
instance of Fullnode. An instance of Fullnode will be lowercased: fullnode. It
is better to uppercase the classname now so that we don't have weird problems
later on.