MsgAlert
Satoshi added this message type in case there were ever a need to issue an
alert to the bitcoin network, for instance if there were a security
vulnerability that required everyone to update their software. This message
type is mostly unused and will be removed from bitcoin soon. As such, I have
not fully implemented the "alert" data structure. Probably the way to handle
this message if it were ever received would be to record it and note it, but
don't both writing software to parse it until that happens, since it is so
unlikely it will ever be used.