partial BIP68 implementation
BIP68 is relative lock-time, where the sequence number is reinterpreted as
either a length of time or a number of blocks since the input transaction was
confirmed. If the seqnum is lower than that length of time or that number of
blocks, then the transaction is invalid. We can't do a full implementation of
BIP68 yet because we do not have a proper full node and can't access previous
blocks. However, we can implement the bitwise operations to convert and mask
the sequence number in the appropriate way. That is what this new class does.
It will need to be updated in the future once we have full node support to do
the rest of the BIP68 logic.