convert to conventional camelCase
All of Fullnode now uses proper camelCase, with one exception: the name
Fullnode. Other than the name, all classes and variables should be using
correct camelCase. If they aren't, it's a bug. This also includes acronyms,
which are annoyingly not entirely capitalized. That is the only consistent way
to do it. Aes is what was AES, for instance. File names now use a dash to
separate words, which as best I can tell is the standard way to do it
(camelCase filenames is a bad idea, because some filesystems are case
insensitive, causing cross-platform incompatibilities).

Besides the name Fullnode, the Bn class (big number) uses "BN" internally,
although when references elsewhere it is always referred to as Bn. That is
because I had a weird bug when trying to rename BN to Bn inside that class, and
temporarily gave up.

This is a huge breaking change, but it's better to do it now than before our
codebase becomes popular (which theoretically will happen after Yours becomes
popular). Another change to take place soon is switching to the new ES2015
class syntax.