getFailureExplanation
When producing complex scripts, the script interpreter fails, and is often
difficult to debug exactly where the interpreter has failed. This method (and
associated getFailureExplanationString) gives valuable information about where
the interpreter failed, including a string form of the current executing script
(which can be the scriptSig, scriptPubKey, or redeemScript) and the particular
opcode where it failed (both numerically as well as the string form of the
opcode).

This is part of a forward-thinking approach to adding better debug information
to Fullnode. We will want to ultimately have some kind of debug standard in
Fullnode. This is only a first step.