support sending undefined args to workers
It turns out that although the workers were supposed to already support basic
javascript types that undefined was a special case and wasn't work. That is
because JSON.stringify(undefined) does not give the string "undefined" but
rather returns undefined. So you can't JSON.stringify and then JSON.parse
undefined. The answer was to treat undefined as a special case. Fortunately the
workers already supported a notion of "headers" for different types - so I just
had to add another header byte for the case of undefined.