ecies async encrypt and decrypt
ECIES is a standard algorithm to encrypt data to someone using their EC public
key. This is extremely slow because of multiple rounds of point multiplication
and hashing that has to occur to do this. I've now added async versions of both
encrypting and decrypting so it can be used in a UI where responsiveness
matters.