allow forcing TxBuilder to use all inputs
The TxBuilder does not use all inputs by default, because it may not be nessary
to use all inputs to spend the amount desired. Sometimes, however, you may want
to force the TxBuilder to use all inputs, such as when you are spending the
full balance of a wallet. The new option useAllInputs allows you to force the
TxBuilder to use all inputs.