break up tests for consistent parallel performance
Individual tests that take a long time should be broken up into separate
files so the parallel test runner can run them in different processes.
I've broken up the slower tests, the Interp and TxVerifier test vectors,
into multiple files so that they are run in separate processes most of
the time. I've also changed the default 'npm test' not to use gulp but
to use mocha directly. The different speed results on my computer are:

- gulp: 17s
- mocha: 6.8s
- test-parallel: 3.4s

gulp is by far the slowest, and I'm not sure why. test-parallel is
definitely the fastest, although it is interesting to note that even on
my 4-core machine, it is only twice as fast and not four times as fast
as series mocha. That is probably because the javascript VM has less
fewer function calls per process to optimize each function.