fix op return
We were unable to pass multiple chunks of data in an OP_RETURN. This is
now considered standard in Bitcoin Cash which is not the same as Bitcoin
Core. ow we can. The problem was twofold. First, we needed to fix the
fromASM method to actually use the larger PUSHDATA opcodes correctly so
that the isDataOut method would work properly. Secondly, we needed to
fix the isPushOnly method to support the larger PUSHDATA opcodes.