new version
- Fix transaction tests due to changed fee

- Remove phantomjs tests - they are broken due to a javascript syntax it
doesn't understand, but it is extremely outdated and I don't want to
bother to fix it. Better to just remove it. We do not need to test for
the browser beacuse we are not doing anything unusual like IndexedDB or
anything else specific to browsers, nor are we doing anything specific
to node. If there is anything wrong with our browser code, then webpack
will fail to build.

- New version number

- Fix scripts to all rely on yarn and webpack. No gulp.