xpriv: to/from Hex (#60)
* "hex buffer" makes no sense. it's just a buffer.

"hex" refers to a string encoding.

* fix test

* shudder

For some reason the to/from Buffer methods simply convert the string
into a buffer, even though an xpub *is* a buffer. In other words, the
characters themselves are actually encoded into the buffer. The way it
should work is that the base58check buffer is what is the "toBuffer". Oh
well -- it's too difficult to change. We're stuck with it now. I have
added to/from Hex methods to be consistent with the rest of the library.
Over time I will add to/from Buffer and to/from Hex methods to all
objects. Additionally, each object will have to/from String and to/from
Object and to/from JSON.

* lint