Fix JSON test vectors: BIP32 checksum value should be unsigned. (#128)
BufferUtil methods integerAsBuffer() and integerFromBuffer() do not
consider the sign of the value, so test vectors with a negative
checksum are accepted.  However, future refactoring of BufferUtil
to use methods such as Buffer.readUInt32BE() will result in a
range exception because of the negative checksum.  To resolve this
issue, the signed checksum value is replaced with its unsigned value.