remove support for P2SH from Address and TxBuilder
It is no longer possible to produce P2SH outputs. See this announcement:

https://bitcoinsv.io/2020/01/15/changes-for-the-genesis-upgrade/

The only wallets that still need P2SH are legacy wallets that need to recover
their funds and send all funds to a new wallet. Those wallets can use the old
version of the library to create a custom one-off tool to recover funds.

Note that in order be able to *validate* transactions, we still need to keep a
notion of P2SH around, because someone may send you a transaction that spends a
P2SH output. Also, we copy the script interpreter from the node software, so it
is important that the validator part of the code stays as similar as possible
to the node software. Thus, we will probably have to keep that part of the code
around forever, unless we eventually learn that all P2SH outputs have been
spent (unlikely).

Address is similar in that because we cannot actually use P2SH addresses
anymore, why do we need that legacy code hanging around? If someone wants to
create some legacy viewer for P2SH addresses, they can do so. But you do not
need to display the address. Just display the script. You can use the methods
on Script to know whether it is P2SH or not and display an indicator - 'legacy
P2SH script found'.