elliptic -> bitcoin-elliptic
The elliptic package by Fedor Indutny is a fantastic cryptography library for
elliptic curve-related methods. However, because we are building for browsers,
we care about file size, and the circular dependency-nature of elliptic
includes a bunch of curves unrelated to Bitcoin and therefore bloating the size
in a browser. Here, we update to bitcoin-elliptic which is the same as elliptic
but with all curves unrelated to Bitcoin removed, making the file size smaller
in a browser leading to faster load times.