rename sign() to signWithKeyPairs()
The interface for the library is in flux while it is in beta. Since there are
already multiple ways to sign a transaction, it makes sense to be explicit
about the way each method does the signing. There many be other ways to sign
transactions in the future. By giving this an explicit name, we can reserve the
name "sign" for some fully generic method in the future, or just slowly add
other explicit ways to sign transactions over time and never have a generic
"sign" method.