Limit range of privkeys from arbitrary buffers
The fromBuffer method assumes that the specified buffer always results
in a valid big number, smaller than the order of the curve, but this is not
always the case with an arbitrary 32-byte array.

Essentially the fromBuffer method allows reconstructing privkeys that
don't pass the validate() check.

This is important to prevent people from creating invalid privkeys.